added productCard
